Added: -MixAndMatchClass -Changed OutputManager class int to float TODO: -Ensure the drag sequence works. -collides and it will transition to this screen.

+++ desktop/src/com/mygdx/game/Managers/OutputManager.java
@@ -21,7 +21,7 @@ public class OutputManager {
     }
 	
 	//Draw Render
-	public void draw(SpriteBatch batch,Texture tex, int x, int y) {
+	public void draw(SpriteBatch batch,Texture tex, float x, float y) {
 		batch.draw(tex,x,y,Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
 	}
